# task1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Films endpoints",
        "title": "Test task",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:28000",
    "basePath": "/",
    "paths": {
        "/film": {
            "get": {
                "description": "Get all films",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "film"
                ],
                "summary": "Get Films",
                "parameters": [
                    {
                        "enum": [
                            "releaseDate",
                            "-releaseDate"
                        ],
                        "type": "string",
                        "description": "sort order",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Add new film",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "film"
                ],
                "summary": "AddFilm",
                "parameters": [
                    {
                        "description": "Film",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/testGolang.Film"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/film/{ID}": {
            "get": {
                "description": "Get Film by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "film"
                ],
                "summary": "Get Film",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Film ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "testGolang.Film": {
            "type": "object",
            "required": [
                "releaseDate",
                "title"
            ],
            "properties": {
                "releaseDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}